backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.28.0
pydantic>=2.7.0
python-dotenv==1.0.1
